RDKAAMP-3962 : Sync subtitles at 1s interval
Reason for change: Video freeze was seen to put subtitles ahead of AV.
We need to more agressively sync them up. This is an interim approach
until we have a PTS progress monitor that could trigger the variable
update rate as needed.

Test Procedure: Check subtitles stay in sync when we introduce 404 errors
or create underflow and video freezes in other ways.
Risks: Low
Priority: P1

Change-Id: Iadd1aaa48867c33d0db234b35799c69d03827f74
Signed-off-by: gavin.parry2@s.ky.uk