VPLAY-9208: [L1] New test case for SelectSourceOrAdPeriod
Reason for change: Test case for SelectSourceOrAdPeriod where the ad in the landing period is not yet resolved,
but player is in IN_ADBREAK_WAIT2CATCHUP
Risks: Low
Test Procedure: Test with
Priority: P1

Change-Id: I5130e5ae15cf09b7d0f069a9f3276539a5dd61c0
Signed-off-by: Nandakishor Udiyannur<Nandakishor_udiyannurmana@comcast.com>

VPLAY-8860: ads played out on the 1st play but same spot got substituted by DAI on rewind

Reason for Change: CDAI ad is not being delivered soon enough to be
used on 1st play, but it does arrive and is used on 2nd play. This
change to invalidate the AD if it is not used on 1st play hence
it will not be used in subsequent play.

Test Procedure. See jira

Risk:Medium

Change-Id: Ia627d818885f9de2d9e3236e3fa685c1d022c62e
Signed-off-by: Nandakishor Udiyannur<Nandakishor_udiyannurmana@comcast.com>