VPLAY-9327 WPEWebProcess crash with fingerprint 76741189
Reason for change:
WebProcess crashes in AampCache::Remove due to assertion `iter2 != cache.end()'

This is because 2 different URLs resolve to the same effective URL. When
the first is removed from the cache the associated effective URL is also
removed. As a result when the second URL is removed the associated
effective URL no longer exists.

This change checks if any other URL reference the effetive URL before
removing it.

Test Procedure: Run L1 tests, and confirm at full stack the issue is
resolved.

Risks: Low

Signed-off-by: Dominic Holden <doholden@synamedia.com>

Change-Id: Ia63449cf82ceb22e6223643b3d6b02841bdeb977