VPLAY-9393 HLS bitrate remains low post stream freeze
Reason for change: When the network bitrate drops low enough that
segments are dropped, when it recovers the bitrate doesn't always ramp
back up to the correct level.

There are 2 issues causing this:
1) The last profile index used could erroneously be restored, leaving
   AAMP playing the low bitrate playlist, but internally thinking it is
   playing a higher bitrate.
2) When segments are dropped the buffer depth calculation will be out by
   number of segments dropped * segment duration. This can then result
   in a -ve buffer depth calculation resulting in staying at a lower
   bitrate than can be supported.

Test procedure: Play HLS content and lower the NW bandwidth such that
the stream freezes, on restoring the NW bandwidth ensure AAMP ramps up
to the correct video quality. (I verified this from logs)

Risk: Low

Signed-off-by: Dominic Holden<doholden@synamedia.com>

Change-Id: Id4f2cb56b8f46963e3f9d251f00544744d888bb4