VPLAY-9520  With AAMP TSB network issues cause return to home screen
Reason for change: With local AAMP TSB enabled if there is underflow the
AV will freeze regardless of if the NW recovers.

This is due to the logic that was added to improve pause, as the
segments are discarded after adding them to TSB in the belief that
playback will resume from TSB after a pause. When there is underflow,
the pipeline is also paused and this scenario was missed.

Test procedure: With Local TSB enabled tune to BBC1 and allow the device
to reach steady state. Lower the NW bandwidth to a point the requests
time out (10KB/s worked nicely for me). Once the AV freezes restore the
NW bandwidth and check AV resumes correctly.

Regression Test: Perform the same test as above but without the Local
TSB enabled and ensure the behaviour is unchanged.

Risk: LOW

Signed-off-by: Dominic Holden<doholden@synamedia.com>

Change-Id: I32c60946d4be89b9279629627c50f69af3b760bf