VPLAY-9566 [ENT-4082] Audio track change after FF to live plays from live edge instead of live playing position
Reason for change: When changing the prefered language whilst injecting
from the local TSB the wrong seek is being used. Resulting in seeking to
the very live edge, and not to the live position.

Test procedure: On a SLD channel with multiple languages (i.e 172) with fog dissabled, and local TSB
enabled:
 Pause for a couple of minutes
 FFWD to live edge so the automatic transition to play occurs
 Change the language from Welsh to English
 Observe play continues, there will be a small AV glitch as we are
effectively seeking to the same position as we are currently playing

Risk: Low

Signed-off-by: Dominic Holden<doholden@synamedia.com>

Change-Id: Id42d7c9bd9a98f913180e7ba77131e3df7762be1